:hammer: Reorganize `omisego-admin` live tests
