+++ omisego-admin/src/liveTest/kotlin/co/omisego/omisego/BaseLiveTest.kt
@@ -11,8 +11,8 @@ import co.omisego.omisego.model.AdminConfiguration
 import co.omisego.omisego.network.ewallet.EWalletAdmin
 import co.omisego.omisego.utils.ResourceFileLoader
 
-open class LiveTest : ResourceFileLoader() {
-    private val secret by lazy { loadSecretFile("secret.json") }
+open class BaseLiveTest : ResourceFileLoader() {
+    val secret by lazy { loadSecretFile("secret.json") }
     private val config by lazy {
         AdminConfiguration(
             secret.getString("base_url")
